requirements.txt
apache-beam==2.43.0
apache-flink==1.17.0
apache-flink-libraries==1.17.0
avro-python3==1.9.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
cloudpickle==2.2.0
confluent-kafka==2.1.0
crcmod==1.7
dill==0.3.1.1
docopt==0.6.2
fastavro==1.4.7
fasteners==0.18
find-libpython==0.3.0
grpcio==1.54.0
hdfs==2.7.0
httplib2==0.20.4
idna==3.4
kafka-python==2.0.2
numpy==1.21.6
objsize==0.5.2
orjson==3.8.11
pandas==1.3.5
pemja==0.3.0
proto-plus==1.22.2
protobuf==3.20.3
py4j==0.10.9.7
pyarrow==8.0.0
pydot==1.4.2
pymongo==3.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
regex==2023.3.23
requests==2.29.0
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.15
websocket-client==1.5.1
zstandard==0.21.0